README Clarification
I mistook the web browser's developer console to mean the server's. Because below in the FAQ it was mentioned on how to tell that it's up by pinging it, so that was the only context I had for using these commands.

This might help other's with the same mindset as me.